Add a Cortex-M0+ bootloader application.
Also changes the psoc6-cm4-rtic application so it can be started by the bootloader.

For some reason, when the CM4 binary runs, the VTOR has not been set the value we put in the cm4_vector_table_base register, so we manually have to patch it once the CM4 is running.

Also when the CM0+ application starts, the stack pointer is set to the stop of SRAM, not the value we put in the vector table - I guess the ROM loader mostly ignores the vector table? Or I don't understand how the ROM loader works. So anyway we have to manually fix that up as well.

All the hprintln! stuff is commented out so it works even if a debugger is not connected.